move functions supporting images command to libkpod/image
