+++ cmd/kpod/common.go
@@ -1,12 +1,8 @@
 package main
 
 import (
-	"strings"
-
 	is "github.com/containers/image/storage"
-	"github.com/containers/image/types"
 	"github.com/containers/storage"
-	"github.com/pkg/errors"
 	"github.com/urfave/cli"
 )
 
@@ -34,19 +30,4 @@ func getStore(c *cli.Context) (storage.Store, error) {
 	}
 	is.Transport.SetStore(store)
 	return store, nil
-}
-
-func parseRegistryCreds(creds string) (*types.DockerAuthConfig, error) {
-	if creds == "" {
-		return nil, errors.New("no credentials supplied")
-	}
-	if strings.Index(creds, ":") < 0 {
-		return nil, errors.New("user name supplied, but no password supplied")
-	}
-	v := strings.SplitN(creds, ":", 2)
-	cfg := &types.DockerAuthConfig{
-		Username: v[0],
-		Password: v[1],
-	}
-	return cfg, nil
 }
